requirements.txt
alembic==1.4.1
aniso8601==8.0.0
astroid==2.3.3
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
pylint==2.4.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
